saving...

+++ src/elt_pipeline/load/load_to_rds.py
@@ -92,7 +92,7 @@ def load_data_into_database(games_df: pd.DataFrame,
     publisher_assignment_df = data["publisher_assignment"]
 
     games_df["app_id"] = games_df["app_id"].astype(int)
-    games_df["price"] = games_df["price"].astype(float)
+    games_df["price"] = games_df["price"].astype(int)
 
     engine = get_engine()
